requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
psycopg[binary]==3.2.3
pydantic==2.10.3
python-dotenv==1.0.1